refactor: :recycle: moved reducer file
